ref: Organize consumer errors in a module
