+++ snuba/subscriptions/consumer.py
@@ -12,7 +12,8 @@ from typing import (
     Sequence,
 )
 
-from streaming_kafka_consumer import Consumer, ConsumerError, Message, Partition, Topic
+from streaming_kafka_consumer import Consumer, Message, Partition, Topic
+from streaming_kafka_consumer.errors import ConsumerError
 
 from snuba.utils.types import Interval, InvalidRangeError
