+++ streaming_kafka_consumer/__init__.py
@@ -1,17 +1,5 @@
-from streaming_kafka_consumer.backends.abstract import (
-    Consumer,
-    ConsumerError,
-    EndOfPartition,
-    Producer,
-)
+from streaming_kafka_consumer import errors
+from streaming_kafka_consumer.backends.abstract import Consumer, Producer
 from streaming_kafka_consumer.types import Message, Partition, Topic
 
-__all__ = [
-    "Consumer",
-    "ConsumerError",
-    "EndOfPartition",
-    "Message",
-    "Partition",
-    "Producer",
-    "Topic",
-]
+__all__ = ["Consumer", "Message", "Partition", "Producer", "Topic", "errors"]
