+++ streaming_kafka_consumer/backends/local/storages/file.py
@@ -16,7 +16,6 @@ from typing import (
     Tuple,
 )
 
-from streaming_kafka_consumer.backends.abstract import OffsetOutOfRange
 from streaming_kafka_consumer.backends.local.storages.abstract import (
     MessageStorage,
     PartitionDoesNotExist,
@@ -24,6 +23,7 @@ from streaming_kafka_consumer.backends.local.storages.abstract import (
     TopicExists,
 )
 from streaming_kafka_consumer.codecs import Codec
+from streaming_kafka_consumer.errors import OffsetOutOfRange
 from streaming_kafka_consumer.types import Message, Partition, Topic, TPayload
