+++ streaming_kafka_consumer/tests/backends/mixins.py
@@ -6,12 +6,11 @@ from typing import ContextManager, Generic, Iterator, Mapping, Optional, Sequenc
 from unittest import mock
 
 import pytest
-from streaming_kafka_consumer.backends.abstract import (
-    Consumer,
+from streaming_kafka_consumer.backends.abstract import Consumer, Producer
+from streaming_kafka_consumer.errors import (
     ConsumerError,
     EndOfPartition,
     OffsetOutOfRange,
-    Producer,
 )
 from streaming_kafka_consumer.tests.assertions import (
     assert_changes,
